app/streamData/consumer/requirements.txt
pandas
kafka-python
pymongo